+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 2.2.8
 * Added global mod registry, collecting added entities per mod
 * Fixed sprite atlas loading
+* Show color on the ColorPicker button in the mod settings
 
 ## Version 2.2.7
 * Added loading of dll embedded text assets (thx MSchmoecker)